requirements.txt
pytest-html-reporter
allure-pytest
